+++ src/index/ix_index_handle.h
@@ -29,8 +29,8 @@ inline int ix_compare(const char* a, const char* b, ColType type, int col_len) {
             return (ia < ib) ? -1 : ((ia > ib) ? 1 : 0);
         }
         case TYPE_FLOAT: {
-            double fa = *(double*)a;
-            double fb = *(double*)b;
+            float fa = *(float*)a;
+            float fb = *(float*)b;
             return (fa < fb) ? -1 : ((fa > fb) ? 1 : 0);
         }
         case TYPE_STRING:
